Add a way to enforce me setting the right user on git commits - fixes #21
This allows me to work like before, but requires chosing the right shortcut when cloning or creating a new repo.
This is totally fine though, just means that all the repos that were using global config will need to `git clux` again.
But this is fine, because email (none) causes commit to fail.